Move SlurmSettings to the Cluster section
